DiscordBots/AESO/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
discord.py==1.7.3
idna==3.3
libretranslatepy==2.1.1
lxml==4.7.1
multidict==5.2.0
requests==2.27.1
translate==3.6.1
typing-extensions==4.0.1
urllib3==1.26.8
yarl==1.7.2